ユーザー一覧を分離する
